feature/R-125
네이밍 수정
