+++ app/src/main/java/com/swsnack/catchhouse/data/db/AppDatabase.java
@@ -3,7 +3,7 @@ package com.swsnack.catchhouse.data.db;
 import com.swsnack.catchhouse.AppApplication;
 import com.swsnack.catchhouse.repository.room.local.RoomDao;
 import com.swsnack.catchhouse.repository.room.local.TypeConverter;
-import com.swsnack.catchhouse.data.entity.RoomEntity;
+import com.swsnack.catchhouse.data.entity.FavoriteRoomEntity;
 
 import androidx.room.Database;
 import androidx.room.Room;
@@ -12,7 +12,7 @@ import androidx.room.TypeConverters;
 
 import static com.swsnack.catchhouse.Constant.DatabaseKey.DATABASE_NAME;
 
-@Database(entities = {RoomEntity.class}, version =  1, exportSchema = false)
+@Database(entities = {FavoriteRoomEntity.class}, version =  1, exportSchema = false)
 @TypeConverters({TypeConverter.class})
 public abstract class AppDatabase extends RoomDatabase {
