+++ app/src/main/java/com/swsnack/catchhouse/util/DataConverter.java
@@ -2,7 +2,7 @@ package com.swsnack.catchhouse.util;
 
 import android.graphics.Bitmap;
 
-import com.swsnack.catchhouse.data.entity.RoomEntity;
+import com.swsnack.catchhouse.data.entity.FavoriteRoomEntity;
 import com.swsnack.catchhouse.data.model.Chatting;
 import com.swsnack.catchhouse.data.model.Message;
 import com.swsnack.catchhouse.data.model.Room;
@@ -47,8 +47,8 @@ public class DataConverter {
         return chattingList;
     }
 
-    public static RoomEntity convertToRoomEntity(Room room) {
-        return new RoomEntity(room.getKey(),
+    public static FavoriteRoomEntity convertToRoomEntity(Room room) {
+        return new FavoriteRoomEntity(room.getKey(),
                 room.getPrice(),
                 room.getFrom(),
                 room.getTo(),
@@ -68,24 +68,24 @@ public class DataConverter {
                 room.isDeleted());
     }
 
-    public static Room convertToRoom(RoomEntity roomEntity) {
-        return new Room(roomEntity.getRoomUid(),
-                roomEntity.getPrice(),
-                roomEntity.getFrom(),
-                roomEntity.getTo(),
-                roomEntity.getTitle(),
-                roomEntity.getContent(),
-                roomEntity.getImages(),
-                roomEntity.getUuid(),
-                roomEntity.getAddress(),
-                roomEntity.getAddressName(),
-                roomEntity.getSize(),
-                roomEntity.isOptionStandard(),
-                roomEntity.isOptionGender(),
-                roomEntity.isOptionPet(),
-                roomEntity.isOptionSmoking(),
-                roomEntity.getLatitude(),
-                roomEntity.getLongitude(),
-                roomEntity.isDeleted());
+    public static Room convertToRoom(FavoriteRoomEntity favoriteRoomEntity) {
+        return new Room(favoriteRoomEntity.getRoomUid(),
+                favoriteRoomEntity.getPrice(),
+                favoriteRoomEntity.getFrom(),
+                favoriteRoomEntity.getTo(),
+                favoriteRoomEntity.getTitle(),
+                favoriteRoomEntity.getContent(),
+                favoriteRoomEntity.getImages(),
+                favoriteRoomEntity.getUuid(),
+                favoriteRoomEntity.getAddress(),
+                favoriteRoomEntity.getAddressName(),
+                favoriteRoomEntity.getSize(),
+                favoriteRoomEntity.isOptionStandard(),
+                favoriteRoomEntity.isOptionGender(),
+                favoriteRoomEntity.isOptionPet(),
+                favoriteRoomEntity.isOptionSmoking(),
+                favoriteRoomEntity.getLatitude(),
+                favoriteRoomEntity.getLongitude(),
+                favoriteRoomEntity.isDeleted());
     }
 }
